backend requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0  # For JWT
passlib[bcrypt]==1.7.4  # Password hashing
reportlab==4.0.7  # PDF reports
AWSIoTPythonSDK==1.5.1  # For MQTT